renamed files to lower case
