+++ src/geouned/GEOUNED/__init__.py
@@ -1,2 +1,2 @@
 from .core import CadToCsg
-from .Utils.data_classes import NumericFormat, Options, Settings, Tolerances
+from .utils.data_classes import NumericFormat, Options, Settings, Tolerances
