+++ src/geouned/GEOUNED/write/mcnp_format.py
@@ -8,10 +8,10 @@ from pathlib import Path
 
 import FreeCAD
 
-from ..CodeVersion import *
-from ..Utils.BasicFunctions_part1 import is_opposite, points_to_coeffs
-from ..Utils.Functions import SurfacesDict
-from .Functions import CardLine, change_surf_sign, mcnp_surface, write_mcnp_cell_def
+from ..code_version import *
+from ..utils.basic_functions_part1 import is_opposite, points_to_coeffs
+from ..utils.functions import SurfacesDict
+from .functions import CardLine, change_surf_sign, mcnp_surface, write_mcnp_cell_def
 
 logger = logging.getLogger("general_logger")
