+++ src/geouned/GEOUNED/write/serpent_format.py
@@ -7,10 +7,10 @@ from pathlib import Path
 
 import FreeCAD
 
-from ..CodeVersion import *
-from ..Utils.BasicFunctions_part1 import is_opposite, points_to_coeffs
-from ..Utils.Functions import SurfacesDict
-from .Functions import change_surf_sign, serpent_surface, write_serpent_cell_def
+from ..code_version import *
+from ..utils.basic_functions_part1 import is_opposite, points_to_coeffs
+from ..utils.functions import SurfacesDict
+from .functions import change_surf_sign, serpent_surface, write_serpent_cell_def
 
 logger = logging.getLogger("general_logger")
