+++ src/geouned/GEOUNED/write/string_functions.py
